requirements.txt
streamlit-ext
streamlit-player
streamlit-extras
plotly
pandas
numpy
numpy-financial
kaleido